Change Satellite's el, az, ss members to be floats
These are floats (e.g., `35.0` for `el`) in the JSON that comes from my
system's `gpsd`.

+++ src/lib.rs
@@ -290,11 +290,11 @@ pub struct Satellite {
     #[serde(rename = "PRN")]
     pub prn: i16,
     /// Elevation in degrees.
-    pub el: i16,
+    pub el: f32,
     /// Azimuth, degrees from true north.
-    pub az: i16,
+    pub az: f32,
     /// Signal strength in dB.
-    pub ss: i16,
+    pub ss: f32,
     /// Used in current solution? (SBAS/WAAS/EGNOS satellites may be
     /// flagged used if the solution has corrections from them, but
     /// not all drivers make this information available.).
